Implement feature to select and 'redo' members
We do have the case quite often, that we want to let a team member go first (if they have to attend another meeting, etc.). With this PR this can be done by clicking on that team members name instead of the 'spotlight'.
A second case that we had sometimes is that a team member was accidentally taken out of the 'spotlight'. With this PR that team member can be put from the 'Done'-list back into the spotlight by clicking on their name. The member currently in the 'spotlight' will be put back in the 'ToDo'-list.