Rewrite app based on yurique/frontroute-example
This allows to use basic routing in the app. This is mainly/only used to replace the hard coded team member list with a list provided in the URL.
If the list is not provided the app will show a form field to generate the URL.
The app also makes use of webpack-dev-server and proper bundling now.